% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.ctd}
\alias{fortify.ctd}
\alias{fortify.section}
\alias{fortify.adp}
\title{Prepare 'oce' objects as data frames for 'ggplot2'}
\usage{
\method{fortify}{ctd}(model, ..., which = c("combined", "data", "metadata"))

\method{fortify}{section}(model, ..., which = c("combined", "data", "metadata"))

\method{fortify}{adp}(model, ..., which = c("velocity", "bottom_track", "metadata"))
}
\arguments{
\item{model}{An 'oce' object}

\item{...}{Additional columns/values to initialize. Tidy evaluation
is supported.}

\item{which}{Most objects can be summarised as data frames at several
levels of granularity. The default is the most useful for plotting.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Prepare 'oce' objects as data frames for 'ggplot2'
}
\examples{
library(ggplot2)
data(ctd, package = "oce")
fortify(ctd)

}
