% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-isopycnal.R
\docType{data}
\name{geom_isopycnal}
\alias{geom_isopycnal}
\alias{GeomIsopycnal}
\title{Draw automagic isopycnals}
\format{
An object of class \code{GeomIsopycnal} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
geom_isopycnal(
  mapping = NULL,
  data = NULL,
  ...,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  colour = "darkgray",
  size = 0.4,
  linetype = 1,
  alpha = NA,
  text.size = 3,
  family = "",
  salinity_type = c("practical", "absolute"),
  temperature_type = c("in-situ", "conservative"),
  ref_pressure = 0,
  ref_longitude = NULL,
  ref_latitude = NULL,
  trim_freezing = TRUE,
  breaks = pretty,
  labels = identity,
  label_placer = NULL,
  n_breaks = 5,
  n_sal = 200,
  n_temp = 200,
  eos = getOption("oceEOS", default = "gsw"),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

GeomIsopycnal
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{lineend, linejoin, linemitre, colour, size, linetype, alpha}{Customize
the appearance of contour lines. See \code{\link[ggplot2:geom_path]{ggplot2::geom_path()}} for details.}

\item{text.size, family}{Customize the appearance of contour line labels.
See \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} for details.}

\item{salinity_type}{A salinity value. Use
\code{salinity_type = "absolute"} to specify that \code{salinity} values
are absolute salinities.}

\item{temperature_type}{A temperature value. Use
\code{temperature_type = "conservative"}
to specify that these values are not in-situ temperature.}

\item{ref_pressure}{The pressure that should be used to calculate
density contours in dbar.}

\item{ref_longitude}{The latitude and longitude that should
be used in seawater calculations if \code{eos} is "gsw".}

\item{ref_latitude}{The latitude and longitude that should
be used in seawater calculations if \code{eos} is "gsw".}

\item{trim_freezing}{Don't draw contours outside the freezing region.}

\item{breaks}{A numeric vector or function used to generate breaks
from objects.}

\item{labels}{A function or vector of labels with which isopycnals
should be labeled.}

\item{label_placer}{See \code{\link[=label_placer_isopycnal]{label_placer_isopycnal()}} and
\code{\link[isoband:label_placer]{isoband::label_placer_minmax()}} for ways to customize
label placement. Use \code{NULL} for a default that avoids data passed
to the layer.}

\item{n_breaks}{Passed to \code{breaks} if \code{breaks} is a function.}

\item{n_sal}{Number of points with which contours should be approximated
on the salinity axis.}

\item{n_temp}{Number of points with which contours should be approximated
on the temperature axis.}

\item{eos}{The equation of state: one of "unesco" or "gsw". See
\code{\link[oce:swRho]{oce::swRho()}} for details.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
A \code{\link[ggplot2:layer]{ggplot2::layer()}}
}
\description{
Draws density contour lines ("isopycnals") as a ggplot2 layer. The
\code{x} axis is assumed to be salinity; the \code{y} axis is assumed to be
temperature. To use this geometry as standalone layer, you will need
to specify which type of temperature (i.e., in-situ or conservative)
and salinity (i.e., practical or absolute) are on these axes.
}
\examples{
library(ggplot2)
data(ctd, package = "oce")

ggplot(ctd, aes(salinity, temperature)) +
  geom_isopycnal() +
  geom_point()

}
\keyword{datasets}
