% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-isopycnal.R
\name{label_placer_isopycnal}
\alias{label_placer_isopycnal}
\title{Isopycnal label placement}
\usage{
label_placer_isopycnal(
  placement = "t",
  padding = c(0.025, 0.025),
  bounds = c(padding[1], padding[2], 1 - padding[1], 1 - padding[2])
)
}
\arguments{
\item{placement}{A string containing one or more of "t", "b", "l",
or "r". See \code{\link[isoband:label_placer]{isoband::label_placer_minmax()}}.}

\item{padding}{A proportion of the plot (x, y) that should be excluded
on all sides when choosing label center points. Ignored if
\code{bounds} is specified.}

\item{bounds}{A vector of xmin, ymin, xmax, ymax, outside which
label centers should not exist.}
}
\value{
A function suitable as input to \code{\link[=geom_isopycnal]{geom_isopycnal()}}'s
\code{label_placer} argument.
}
\description{
Automatic labels placement is difficult and infrequently perfect. The
default strategy to place labels uses a similar strategy to
\code{\link[isoband:label_placer]{isoband::label_placer_minmax()}} except coordinates are constrained
to a fraction of the visible area to keep labels from disappearing
outside the plot bounds.
}
