% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-oce.R
\name{theme_oce}
\alias{theme_oce}
\title{Base theme useful for 'oce' plots}
\usage{
theme_oce(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw()}}}
}
\value{
A complete \code{\link[ggplot2:theme]{ggplot2::theme()}}
}
\description{
Base theme useful for 'oce' plots
}
\examples{
library(ggplot2)
ggplot(mpg, aes(hwy, cty)) +
  geom_point() +
  theme_oce()

}
