% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.ctd}
\alias{fortify.ctd}
\alias{fortify.section}
\title{Prepare 'oce' objects as data frames for 'ggplot2'}
\usage{
\method{fortify}{ctd}(model, ...)

\method{fortify}{section}(model, ...)
}
\arguments{
\item{model}{An 'oce' object}

\item{...}{Additional columns/values to initialize.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Prepare 'oce' objects as data frames for 'ggplot2'
}
\examples{
library(ggplot2)
data(ctd, package = "oce")
fortify(ctd)

}
