% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-isopycnal.R
\name{label_placer_isopycnal}
\alias{label_placer_isopycnal}
\title{Isopycnal label placement}
\usage{
label_placer_isopycnal(bounds = c(0.1, 0.1, 0.9, 0.9), data_avoid = NULL)
}
\arguments{
\item{bounds}{A vector of xmin, ymin, xmax, ymax, outside which
label centers should not exist.}

\item{data_avoid}{An list of x, y values that the placer should
attempt to avoid when choosing a label center.}
}
\value{
A function suitable as input to \code{\link[=geom_isopycnal]{geom_isopycnal()}}'s
\code{label_placer} argument.
}
\description{
Automatic labels placement is difficult and infrequently perfect. The
default strategy to place labels uses a similar strategy to
\code{\link[isoband:label_placer]{isoband::label_placer_minmax()}} except coordinates are constrained
to a fraction of the visible area to keep labels from disappearing
outside the plot bounds. Optionally, one can pass in data to "avoid"
while placing the label. Note that at the point the label placer
receives the data, coordinates have been normalized to the range
0-1.
}
